+++ backend/canisters/community/impl/src/updates/mod.rs
@@ -40,7 +40,7 @@ pub mod leave_channel;
 pub mod pin_message;
 pub mod register_poll_vote;
 pub mod register_proposal_vote;
-pub mod register_proposal_vote_v2;
+//pub mod register_proposal_vote_v2;
 pub mod remove_member;
 pub mod remove_member_from_channel;
 pub mod remove_reaction;